+++ src/core/tamper/space2ifs.py
@@ -20,7 +20,7 @@ Replaces space character (' ') with the internal field separator ('$IFS').
 The internal field separator refers to a variable which defines the character 
 or characters used to separate a pattern into tokens for some operations.
 Notes:
-  * This tamper script works against *nix targets
+  * This tamper script works against *nix targets.
 """
 
 if settings.TARGET_OS != "win":